ocrservice/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
numpy==1.24.4
python-dotenv==1.0.0




